New island check function
This file adds the checkislandaccess function and other internal functions.
With this, it is possible to check, if the player is allowed to build on the island. This also allows further changes to later allow other people to build on the island, other than members and island leaders.
And also flags for islands, which allow specific actions are now possible to integrate.