Removed issues with island location determination
In specific cases, there where problems with the determination of an island. By using the location of blocks, we can be sure to have always the right location. Also updated to the new isWithinArea(loc,loc,loc), which checks if the first loc is within the other ones, since the one from SkQuery seems to not work on the current version.