Removed java.lang.Math & use own our functions
java.lang.Math returned rounded numbers sometimes, which is quite unhandy. The new functions getmax and getmin always return the numbers which have been put in without rounding, even, if one of the inputs is an integer and the other is an number, which could be sometimes the case, if a location is 100 and the other is -99.5, with java.lang.Math, it returned -99 instead.